+++ src/core/utils/__tests__/pathHelpers.test.ts
@@ -1,15 +1,15 @@
 import assert from 'node:assert';
 import { describe, test } from 'node:test';
 
-import { relative } from '../pathHelpers';
+import { relativeHelper } from '../../../common/utils/pathHelpers';
 
 describe('path', () => {
     test('@unit: should relative', () => {
-        assert.strictEqual(relative('/test/server', '/test/model'), '../model');
-        assert.strictEqual(relative('/test/server', '/test/server'), './');
-        assert.strictEqual(relative('/test/server', '/test/server/model'), './model');
-        assert.strictEqual(relative('/test/server', '/model'), '../../model');
-        assert.strictEqual(relative('/test', '/test'), './');
-        assert.strictEqual(relative('/test', '/test/model'), './model');
+        assert.strictEqual(relativeHelper('/test/server', '/test/model'), '../model');
+        assert.strictEqual(relativeHelper('/test/server', '/test/server'), './');
+        assert.strictEqual(relativeHelper('/test/server', '/test/server/model'), './model');
+        assert.strictEqual(relativeHelper('/test/server', '/model'), '../../model');
+        assert.strictEqual(relativeHelper('/test', '/test'), './');
+        assert.strictEqual(relativeHelper('/test', '/test/model'), './model');
     });
 });