+++ src/core/utils/precompileTemplates.ts
@@ -1,13 +1,14 @@
 import fs from 'fs';
 import Handlebars from 'handlebars';
-import path from 'path';
+import { extname } from 'path';
 
 import { ELogLevel, ELogOutput } from '../../common/Enums';
 import { Logger } from '../../common/Logger';
-import { SEARCH_REGEXP } from '../../core/types/Consts';
+import { dirNameHelper, joinHelper, relativeHelper, resolveHelper } from '../../common/utils/pathHelpers';
+import { REGEX_BACKSLASH } from '../../core/types/Consts';
 
-const templatesDir = path.resolve(__dirname, '../../templates');
-const compiledDir = path.resolve(__dirname, '../../templatesCompiled');
+const templatesDir = resolveHelper(__dirname, '../../templates');
+const compiledDir = resolveHelper(__dirname, '../../templatesCompiled');
 
 const header = `// This is an automatically generated file for the hbs template.
 // You don't need to change it, run npm run build:hbs to update it.
@@ -37,14 +38,13 @@ const precompileTemplates = () => {
                 if (file.name === '__mocks__') {
                     return;
                 }
-                const filePath = path.join(dir, file.name);
+                const filePath = joinHelper(dir, file.name);
                 if (file.isDirectory()) {
                     walk(filePath);
-                } else if (file.isFile() && path.extname(file.name) === '.hbs') {
-                    const templateName = path
-                        .relative(templatesDir, filePath)
+                } else if (file.isFile() && extname(file.name) === '.hbs') {
+                    const templateName = relativeHelper(templatesDir, filePath)
                         .replace(/\.hbs$/, '')
-                        .replace(SEARCH_REGEXP, '/');
+                        .replace(REGEX_BACKSLASH, '/');
                     const templateContent = fs.readFileSync(filePath, 'utf8');
                     const precompiled = Handlebars.precompile(templateContent, {
                         strict: true,
@@ -58,12 +58,15 @@ const precompileTemplates = () => {
                             union: true,
                             intersection: true,
                             enumerator: true,
+                            containsSystemName: true,
+                            normalizePath: true,
+                            joinPath: true,
                         },
                     });
 
                     const tsContent = header + (precompiled as any);
-                    const outputFilePath = path.join(compiledDir, `${templateName}.ts`);
-                    const outputDir = path.dirname(outputFilePath);
+                    const outputFilePath = joinHelper(compiledDir, `${templateName}.ts`);
+                    const outputDir = dirNameHelper(outputFilePath);
 
                     if (!fs.existsSync(outputDir)) {
                         fs.mkdirSync(outputDir, { recursive: true });